Remove lp dataset file
